+++ apps/drec-api/migrations/1662610170745-CheckcretificatelogforDeviceField.ts
@@ -11,7 +11,7 @@ export class CheckcretificatelogforDeviceField1662610170745 implements Migration
                 "issuer_certificate_id" integer,
                 "certificate_issuance_startdate" timestamp with time zone,
                 "certificate_issuance_enddate" timestamp with time zone,
-                "readvalue_watthour" integer,
+                "readvalue_watthour" DOUBLE PRECISION,
                 "status" character varying,
                 "createdAt" timestamp with time zone NOT NULL DEFAULT now(),
                 "updatedAt" timestamp with time zone NOT NULL DEFAULT now(),
